Cleaned up docs
