+++ orchestrator/src/oldjunk/clear_outputs.tzs
