--- terraform/iam.json
@@ -1,87 +1,0 @@
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Effect": "Allow",
-            "Action": [
-                "ecs:CreateCluster",
-                "ecs:DeleteCluster",
-                "ecs:CreateService",
-                "ecs:DeleteService",
-                "ecs:UpdateService",
-                "ecs:RegisterTaskDefinition",
-                "ecs:DeregisterTaskDefinition",
-                "ecs:ListTaskDefinitions",
-                "ecs:DescribeTaskDefinition",
-                "ecs:PutClusterCapacityProviders",
-                "ecs:DescribeClusters"
-            ],
-            "Resource": "*"
-        },
-        {
-            "Effect": "Allow",
-            "Action": [
-                "iam:CreateRole",
-                "iam:DeleteRole",
-                "iam:GetRole",
-                "iam:PutRolePolicy",
-                "iam:DeleteRolePolicy",
-                "iam:AttachRolePolicy",
-                "iam:DetachRolePolicy",
-                "iam:PassRole"
-            ],
-            "Resource": "arn:aws:iam::*:role/orchestrator-*"
-        },
-        {
-            "Effect": "Allow",
-            "Action": [
-                "secretsmanager:CreateSecret",
-                "secretsmanager:DeleteSecret",
-                "secretsmanager:GetSecretValue",
-                "secretsmanager:PutSecretValue",
-                "secretsmanager:UpdateSecret",
-                "secretsmanager:TagResource"
-            ],
-            "Resource": "arn:aws:secretsmanager:*:*:secret:/orchestrator/*"
-        },
-        {
-            "Effect": "Allow",
-            "Action": [
-                "rds:CreateDBInstance",
-                "rds:DeleteDBInstance",
-                "rds:ModifyDBInstance",
-                "rds:DescribeDBInstances",
-                "rds:CreateDBSubnetGroup",
-                "rds:DeleteDBSubnetGroup",
-                "rds:ModifyDBSubnetGroup",
-                "rds:AddTagsToResource"
-            ],
-            "Resource": "*"
-        },
-        {
-            "Effect": "Allow",
-            "Action": [
-                "logs:CreateLogGroup",
-                "logs:DeleteLogGroup",
-                "logs:PutRetentionPolicy"
-            ],
-            "Resource": "arn:aws:logs:*:*:log-group:*"
-        },
-        {
-            "Effect": "Allow",
-            "Action": [
-                "ec2:CreateSecurityGroup",
-                "ec2:DeleteSecurityGroup",
-                "ec2:AuthorizeSecurityGroupIngress",
-                "ec2:RevokeSecurityGroupIngress",
-                "ec2:CreateVpcEndpoint",
-                "ec2:DeleteVpcEndpoints",
-                "ec2:DescribeVpcEndpoints",
-                "ec2:DescribeSecurityGroups",
-                "ec2:DescribeNetworkInterfaces",
-                "ec2:CreateTags"
-            ],
-            "Resource": "*"
-        }
-    ]
-}
