--- terraform/iam_roles.tf
@@ -1,83 +1,0 @@
-# iam_roles.tf
-
-data "aws_caller_identity" "current" {}
-
-resource "aws_iam_role" "task_role" {
-  name               = "orchestrator-task-role"
-  assume_role_policy = data.aws_iam_policy_document.task_role_policy.json
-}
-
-resource "aws_iam_role" "execution_role" {
-  name               = "orchestrator-execution-role"
-  assume_role_policy = data.aws_iam_policy_document.execution_role_policy.json
-}
-
-data "aws_iam_policy_document" "task_role_policy" {
-  statement {
-    actions = ["sts:AssumeRole"]
-    principals {
-      type        = "Service"
-      identifiers = ["ecs-tasks.amazonaws.com"]
-    }
-  }
-}
-
-data "aws_iam_policy_document" "execution_role_policy" {
-  statement {
-    actions = ["sts:AssumeRole"]
-    principals {
-      type        = "Service"
-      identifiers = ["ecs-tasks.amazonaws.com"]
-    }
-  }
-}
-
-# Attach necessary policies to orchestrator-task-role
-resource "aws_iam_role_policy" "orchestrator_task_policy" {
-  name = "orchestrator-task-policy"
-  role = aws_iam_role.task_role.id
-
-  policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Effect = "Allow"
-        Action = [
-          "ecs:RunTask",
-          "ecs:StopTask",
-          "ecs:DescribeTasks",
-          "ecs:ListTasks",
-          "ecs:DescribeTaskDefinition",
-          "ecs:ListTaskDefinitions",
-          "ecs:DescribeServices",
-          "ecs:ListServices",
-          "ecs:DescribeClusters",
-          "ecs:ListClusters"
-        ]
-        Resource = "*"
-      },
-      {
-        Effect = "Allow"
-        Action = [
-          "iam:PassRole"
-        ]
-        Resource = [
-          aws_iam_role.execution_role.arn,
-          aws_iam_role.task_role.arn
-        ]
-      }
-    ]
-  })
-}
-
-
-# Attach policies to the execution role
-resource "aws_iam_role_policy_attachment" "ecs_task_execution" {
-  role       = aws_iam_role.execution_role.name
-  policy_arn = "arn:aws:iam::aws:policy/service-role/AmazonECSTaskExecutionRolePolicy"
-}
-
-resource "aws_iam_role_policy_attachment" "ecs_logs_policy" {
-  role       = aws_iam_role.execution_role.name
-  policy_arn = "arn:aws:iam::aws:policy/CloudWatchLogsFullAccess"
-}
