--- terraform/locals.tf
@@ -1,23 +1,0 @@
-# Local variables
-
-locals {
-  # Default VPC subnets - you may want to customize these for your environment
-  subnet_ids = data.aws_subnets.default.ids
-}
-
-# Data sources for networking
-data "aws_vpc" "default" {
-  default = true
-}
-
-data "aws_subnets" "default" {
-  filter {
-    name   = "vpc-id"
-    values = [data.aws_vpc.default.id]
-  }
-}
-
-data "aws_security_group" "default" {
-  vpc_id = data.aws_vpc.default.id
-  name   = "default"
-}
