--- terraform/package.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "ao-process-clients": "^2.3.17"
-  },
-  "devDependencies": {
-    "@types/node": "^22.9.1"
-  }
-}
