--- terraform/terraform.tfvars.example
@@ -1,31 +1,0 @@
-# AWS Region
-aws_region = "us-east-1"
-
-# Provider Configuration
-provider_id = "your-provider-id"  # Set this to your unique provider identifier
-
-# Database Configuration (for local development/testing)
-local_db_user     = "myuser"      # Change this
-local_db_password = "mypassword"  # Change this
-db_name           = "orchestrator_db"
-
-# Wallet Configuration
-# Copy the contents of your wallet.json file and paste it here
-local_wallet_json = <<EOT
-{
-  "kty": "RSA",
-  "e": "test",
-  "n": "test",
-  "d": "test",
-  "p": "test",
-  "q": "test",
-  "dp": "test",
-  "dq": "test",
-  "qi": "test"
-}
-EOT
-
-# Secrets Configuration
-secrets_prefix = "/orchestrator"  # Prefix for AWS Secrets Manager
-db_credentials_secret_name = "db-credentials"  # Name for database credentials secret
-wallet_secret_name = "wallet-json"  # Name for wallet secret
